+++ mkchromecast/pulseaudio.py
@@ -14,15 +14,29 @@ def create_sink():
         'module-null-sink',
         'sink_name='+sink_name
         ]
+
     rename_sink = [
         'pacmd',
         'update-sink-proplist',
         sink_name,
         'device.description='+sink_name
         ]
-    subprocess.Popen(create_sink)
+
+    cs = subprocess.Popen(
+            create_sink,
+            stdout=subprocess.PIPE,
+            stderr=subprocess.PIPE
+            )
+    csoutput, cserror = cs.communicate()
+
     time.sleep(1)
-    subprocess.Popen(rename_sink)
+
+    rs = subprocess.Popen(
+            rename_sink,
+            stdout=subprocess.PIPE,
+            stderr=subprocess.PIPE
+            )
+    rsoutput, rserror = rs.communicate()
     return
 
 def remove_sink():
@@ -31,5 +45,11 @@ def remove_sink():
         'unload-module',
         'module-null-sink'
         ]
-    subprocess.Popen(remove_sink)
+
+    rms = subprocess.Popen(
+            remove_sink,
+            stdout=subprocess.PIPE,
+            stderr=subprocess.PIPE
+            )
+    rmsoutput, rmserror = rms.communicate()
     return