+++ changelog.md
@@ -1,3 +1,13 @@
+* mkchromecast (0.3.1) **released**: 2016/08/12
+
+    - Improved Preferences window.
+    - The node streaming has been updated to work with `node v6.3.1`.
+    - Improvements in pulseaudio.py for preventing subprocess.Popen from
+      displaying output.
+    - Improvements in `Check For Updates` method.
+    - Added new option `--chunk-size` to control chunk's size of flask server
+      when streaming using `ffmpeg` or `avconv`.
+
 * mkchromecast (0.3.0) **released**: 2016/07/12
 
     - Youtube URLs can be played piping the audio using `youtube-dl`.